+++ carp_mobile_sensing/lib/runtime/notification/local_notification_controller.dart
@@ -37,6 +37,7 @@ class FlutterLocalNotificationController implements NotificationController {
           onDidReceiveNotificationResponse,
       onDidReceiveNotificationResponse: onDidReceiveNotificationResponse,
     );
+
     info('$runtimeType initialized.');
     debug('PENDING NOTIFICATIONS:');
     for (var notification