+++ carp_mobile_sensing/lib/runtime/notification/notification_controller.dart
@@ -12,14 +12,20 @@ abstract class NotificationController {
   /// The upper limit of scheduled notification on iOS.
   static const PENDING_NOTIFICATION_LIMIT = 64;
 
-  /// The id of the Android notification channel.
+  /// The id of the notification channel.
   static const CHANNEL_ID = 'carp_mobile_sensing_notification_user_tasks';
 
-  /// The name of the Android notification channel as shown in the Settings
+  /// The name of the notification channel as shown in the Settings
   /// on Android phones.
   static const CHANNEL_NAME = 'User Tasks';
 
+  /// The description of the notification channel as shown in the Settings
+  /// on Android phones.
+  static const CHANNEL_DESCRIPTION =
+      'Notifications about tasks that the user has to do.';
+
   /// Initialize and set up the notification controller.
+  /// Also tries to get permissions to send notifications.
   Future<void> initialize();
 
   /// Send an immediate notification for a [task].